clean up unused code
