+++ packages/parser/src/code-edit/index.ts
@@ -1,3 +1,4 @@
+export * from './config';
 export * from './group';
 export * from './image';
 export * from './insert';
@@ -6,4 +7,3 @@ export * from './remove';
 export * from './style';
 export * from './text';
 export * from './transform';
-export * from './framework';
